Layouts : Move keyboard modifier handling into Widget
