+++ python/GafferUI/__init__.py
@@ -289,7 +289,6 @@ from AnimationEditor import AnimationEditor
 import CompoundNumericNoduleUI
 import Examples
 from NameValuePlugValueWidget import NameValuePlugValueWidget
-from Keyboard import Keyboard
 
 # and then specific node uis
